Resolve #1029: Support files and folders to be dragged in externally into project navigator
* Support holding option while dropping to copy
