+++ CodeEdit/Features/NavigatorSidebar/ProjectNavigator/OutlineView/OutlineViewController.swift
@@ -199,7 +199,13 @@ extension OutlineViewController: NSOutlineViewDataSource {
             }
 
             // Needs to come before call to .removeItem or else race condition occurs
-            guard let srcFileItem = try? workspace?.workspaceClient?.getFileItem(fileItemURL.relativePath) else {
+            var srcFileItem: WorkspaceClient.FileItem? = try? workspace?.workspaceClient?.getFileItem(fileItemURL.path)
+            // If srcFileItem is nil, fileItemUrl is an external file url.
+            if srcFileItem == nil {
+                srcFileItem = WorkspaceClient.FileItem(url: URL(fileURLWithPath: fileItemURL.path))
+            }
+
+            guard let srcFileItem else {
                 return false
             }
 
@@ -214,8 +220,11 @@ extension OutlineViewController: NSOutlineViewDataSource {
                     fatalError(error.localizedDescription)
                 }
             }
-
-            self.moveFile(file: srcFileItem, to: destURL)
+            if info.draggingSourceOperationMask == .copy {
+                self.copyFile(file: srcFileItem, to: destURL)
+            } else {
+                self.moveFile(file: srcFileItem, to: destURL)
+            }
         }
         return true
     }
@@ -373,19 +382,5 @@ extension OutlineViewController: NSMenuDelegate {
             }
         }
         menu.update()
-    }
-}
-
-// MARK: - OutlineTableViewCellDelegate
-
-extension OutlineViewController: OutlineTableViewCellDelegate {
-    func moveFile(file: Item, to destination: URL) {
-        if !file.isFolder {
-            workspace?.closeTab(item: .codeEditor(file.id))
-        }
-        file.move(to: destination)
-        if !file.isFolder {
-            workspace?.openTab(item: file)
-        }
     }
 }
